+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.2.0
 
+* Add `stream.fun.writable()` to get a promisey way of asserting a stream is ready for data
 * Add `stream.toJson()` which takes a series of objects and emits them as a JSON stringified array.
 * Add `stream.toNdjson()` which takes a series of objects and emits them as new-line delimited JSON.
 * Add `stream.fromJson()` which is the same as `stream.json()`